requirements.txt
discord.py==2.6.3
requests==2.32.4
Flask==3.1.0
flask-cors==6.0.1